+++ api-admin/controllers/news/review.js
@@ -6,6 +6,7 @@ import { newsLog } from '../../../libs';
 import redis from '../../../redis';
 import moment from 'moment-timezone';
 import libs from '../../../libs';
+import elasticsearch from '../../../elasticsearch';
 
 module.exports = async (req, res, next) => {
     try {
@@ -115,7 +116,7 @@ module.exports = async (req, res, next) => {
         // 處理 log
         updatedNews = await updatedNews.populate('MainMenu Menus MainPhoto MainVideo Photos Videos Author Tags LastReviewer CreatedBy UpdatedBy').execPopulate();
         await newsLog(updatedNews, 'UPDATE');
-
+        await elasticsearch.remove(updatedNews);
         return res.json(updatedNews);
     }catch(err) {
         return next(err);