+++ elasticsearch/index.js
@@ -1,0 +1,12 @@
+import client from './client';
+import create from './create';
+import remove from './remove';
+import search from './search';
+
+
+module.exports = {
+    client,
+    create,
+    remove,
+    search
+};